% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_function_list}
\alias{as_function_list}
\title{Create \code{function_list} object from list of functions}
\usage{
as_function_list(flist, ...)
}
\arguments{
\item{flist}{List in which each entry is a function object. Can be named or unnamed.}

\item{...}{Passed to \verb{\link{function_list}}.}
}
\value{
Function with a single input which outputs a data.frame.
}
\description{
Create a \code{function_list} object from a list of functions. This is useful
if such a list is created programmatically.
}
