% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colapply_functions}
\alias{colapply_functions}
\title{Apply list of functions to column of object}
\usage{
colapply_functions(obj, flist)
}
\arguments{
\item{obj}{2-dimensional array (matrix or data.frame).}

\item{flist}{List of functions of length equal to the number of columns of \code{obj}.
Each entry must be a function applicable to a single column of \code{obj}.
The i-th entry of \code{flist} is applied to the i-th column of \code{obj}.}
}
\value{
Matrix or data.frame (same type as \code{obj}) with names taken from \code{obj}.
}
\description{
Helper function to simplify workflow with lists of functions.
}
