% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_source.R
\name{common_source_model}
\alias{common_source_model}
\title{Model with common sources, but different scores for each}
\usage{
common_source_model(W, S, opts)
}
\arguments{
\item{W}{A list of matrices containing scores for each table.}

\item{S}{A single source matrix underlying all the tables.}
}
\value{
A list of matrices, each with the same sources S but different
scores W.
}
\description{
Model with common sources, but different scores for each
}

