% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{conditional_simulate_data}
\alias{conditional_simulate_data}
\title{Simulate data which satisfies certain conditions}
\usage{
conditional_simulate_data(relations, n_obs, reject = NULL,
  reject_max_iter = 10, on_reject = "ignore", ...)
}
\arguments{
\item{reject}{List of functions. Specifies when a simulated final datamatrix X should
be rejected. Functions must take X as single input and output TRUE if
condition IS NOT met / FALSE if condition IS met and matrix can be
accepted. See details.}

\item{reject_max_iter}{In case of rejection, how many times should a new datamatrix be simulated
until the conditions in \code{reject} are met?}

\item{on_reject}{If "stop", an error is returned if after \code{reject_max_iter} times no
suitable datamatrix X could be found. If "current", the current datamatrix
is returned, regardless of the conditions in \code{reject}.
Otherwise, NULL is returned. In each case a warning is reported.}

\item{...}{All further parameters are passed to \code{\link{simulate_data}}.}
}
\value{
Data.frame with simulated datamatrix X if a suitable dataset can be found
or the iteration limit is hit.
}
\description{
Generate simulated dataset based on transformation of
multivariate gaussian distribution while checking certain
conditions are met.
}
\details{
For details on generating and post-processing datasets, see
\code{\link{simulate_data}}. This function simulates data conditional
on certain requirements that must be met by the final datamatrix X.
This checking is conducted on the output of \code{simulate_data} (i.e.
also includes possible post-processing steps).
}
\note{
Note that \code{relations} specifies the correlation / covariance
of the underlying gaussian data Z and thus does not directly translate into
correlations between the variables of the final datamatrix X.

This function is best used in conjunction with the \code{design} S3 class,
which facilitates further data visualization and conveniently stores
information as a template for simulation tasks.
}
\section{Rejecting Datasets}{

Examples for restrictions include
variance restrictions (e.g. no constant columns which could happen due
to extreme transformations of the initial gaussian distribution Z), ensuring
a sufficient number of observations in a given class (e.g. certain
binary variables should have at least x\% events) or preventing
multicollinearity (e.g. X must have full column rank). If one of the
functions in \code{reject} evaluates to FALSE, the current datamatrix
X is rejected.
In case of rejection, new datasets can be simulated until the conditions
are met or a given maximum iteration limit is hit (\code{reject_max_iter}),
after which the last datamatrix is returned or an error is reported.
}

\section{Rejection Functions}{

Rejection function templates are found in \code{\link{is_collinear}} and
\code{\link{contains_constant}}.
}

\seealso{
\code{\link{design}},
\code{\link{simulate_data}},
\code{\link{is_collinear}},
\code{\link{contains_constant}}
}
