% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utilities.R
\name{contains_constant}
\alias{contains_constant}
\title{Check if matrix contains constant column(s)}
\usage{
contains_constant(x)
}
\arguments{
\item{x}{Matrix or Data.frame.}
}
\value{
TRUE if one of the columns has standard deviation of 0, else FALSE.
}
\description{
Check if matrix contains constant column(s)
}
\note{
Prints a warning if constant is found.
}
