% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cor_from_upper}
\alias{cor_from_upper}
\title{Build correlation matrix}
\usage{
cor_from_upper(n_var, entries = NULL)
}
\arguments{
\item{n_var}{Integer, number of variables (= rows = columns of matrix).}

\item{entries}{Matrix of correlation entries. Consists of 3 columns
(variable_1, variable_2, correlation) that specify both variables
and corresponding correlation in the upper triangular part of
the matrix (i.e. variable_1 < variable_2) .}
}
\value{
Matrix with user supplied entries.
}
\description{
Use to specify correlation matrix in convenient way
by giving entries of the upper triangular part.
}
\examples{
cor_from_upper(2, rbind(c(1,2,0.8)))

}
\seealso{
\code{\link{cor_to_upper}}
}
