% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cor_to_cov}
\alias{cor_to_cov}
\title{Convert correlation matrix to covariance matrix}
\usage{
cor_to_cov(m, sds = NULL)
}
\arguments{
\item{m}{Symmetric correlation matrix.}

\item{sds}{Standard deviations of the variables. Set to 1 for all varirables by default.}
}
\value{
Symmetric covariance matrix.
}
\description{
Rescale correlation matrix by variable standard deviations to yield a
covariance matrix.
}
