% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utilities.R
\name{cor_to_upper}
\alias{cor_to_upper}
\title{Convert correlation matrix to specification used by
\code{cor_from_upper}}
\usage{
cor_to_upper(m)
}
\arguments{
\item{m}{Symmetric correlation matrix.}
}
\value{
Matrix with 3 columns (variable_1, variable_2, correlation), where
correlation gives the entry at position (variable_1, variable_2) of the
input correlation matrix.
}
\description{
Convert correlation matrix to specification used by
\code{cor_from_upper}
}
\seealso{
\code{\link{cor_from_upper}}
}
