% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cov_to_cor}
\alias{cov_to_cor}
\title{Convert covariance matrix to correlation matrix}
\usage{
cov_to_cor(m)
}
\arguments{
\item{m}{Symmetric covariance matrix.}
}
\value{
Symmetric correlation matrix.
}
\description{
Wrapper for \code{\link[stats:cov2cor]{stats::cov2cor}}.
}
\seealso{
\code{\link{cor_to_cov}}
}
