% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{function_list}
\alias{function_list}
\title{Apply list of functions to input}
\usage{
function_list(...)
}
\arguments{
\item{...}{Named or unnamed arguments, each of which is a function taking exactly 
one input. See details.}
}
\value{
Function with a single input.
}
\description{
Apply list of functions to input
}
\details{
This is a convenience function which takes a number of functions and returns
another function which applies all of the user specified functions to a new 
input, and collects the results as data.frame.

Each of the user specified functions is expected to take a single input. 
This is useful to e.g. transform columns of a data.frame. See the example.
}
\note{
This function works fine without naming the input arguments, but the 
resulting data.frames have empty column names if that is the case. Thus, 
it is recommended to only pass named function arguments.
}
\examples{
\dontrun{
f = function_list(
   v1 = function(x) x[,1] * 2, 
   v2 = function(x) x[,2] + 10)
   
f(diag(2))
} 

}
