% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_collinear}
\alias{is_collinear}
\title{Check if matrix is collinear}
\usage{
is_collinear(x)
}
\arguments{
\item{x}{Matrix or Data.frame.}
}
\value{
TRUE if matrix is collinear, else FALSE.
}
\description{
Check if matrix is collinear
}
\note{
Prints a warning if collinear.
}
