% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_cor_matrix}
\alias{is_cor_matrix}
\title{Check if matrix is a correlation matrix}
\usage{
is_cor_matrix(m)
}
\arguments{
\item{m}{Matrix.}
}
\value{
TRUE if matrix is a correlation matrix, else FALSE.
}
\description{
Checks if matrix is numeric, symmetric, has diagonal elements of one,
has only entries in \link{-1, 1}, and is positive definite. Prints a warning
if a problem was found.
}
