% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppca.R
\name{merge_ppca_opts}
\alias{merge_ppca_opts}
\title{Default options for PPCA simulation}
\usage{
merge_ppca_opts(opts)
}
\arguments{
\item{opts}{A (potentially partially specified) list of options to use when
simulating from the probabilistic PCA model. The potential arguments are \cr
   $k_shared The latent dimension for the shared sources. \cr
   $k_unique The latent dimensions for the table-unique sources.
   $n The total number of samples, for each table.
   $p The number of features, for each table.
   $sigma The noise variance.
   $sigma0 The variance when generating scores / factors.}
}
\value{
opts The original opts with defaults filled in.
}
\description{
Default options for PPCA simulation
}

