% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{merge_rsymm_opts}
\alias{merge_rsymm_opts}
\title{Merge default parameters for random symmetric spike generation}
\usage{
merge_rsymm_opts(opts = list())
}
\arguments{
\item{opts}{A (partially filled) list of parameters to use when sampling
spikes. The currently available options are,
  $lambda_m The poisson parameter for the number of spikes to draw
  $lambda_h The exponential parameter for the height of the spikes.
  $lambda_w The exponential parameter for the width of the spikes.
  $alpha_x0 The first shape beta parameter for the positions of the
spikes.
  $beta_x0 The second shape beta parameter for the positions of the
spikes.
  $M The rescaling factor for the positions of the spikes (so they aren't
  just in the unit interval.}
}

