% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdesign.R
\name{mvtnorm_simdesign}
\alias{mvtnorm_simdesign}
\title{Design specification for simulating datasets}
\usage{
mvtnorm_simdesign(relations_initial, mean_initial = 0, sd_initial = 1,
  is_correlation = TRUE, method = "svd",
  name = "Multivariate-normal based simulation design", ...)
}
\arguments{
\item{relations_initial}{Correlation / Covariance matrix of the initial multivariate
gaussian distribution \code{Z}.}

\item{mean_initial}{Vector of mean values of the initial multivariate gaussian
distribution \code{Z}. Dimension needs to correspond to dimension
of \code{relations}.}

\item{sd_initial}{Vector of standard deviations of the initial multivariate
gaussian distribution Z. Dimension needs to correspond to dimension
of \code{relations}. Overriden by suqare root of diagonal elements of
\code{relations} if \code{is_correlation} is FALSE.}

\item{is_correlation}{If TRUE, then \code{relations} specifies a correlation matrix (default,
this type of specification is usually more natural than specifying
a covariance matrix). Otherwise, \code{relations} specifies a
covariance matrix whose square root diagonal elements override
\code{sd_initial}.}

\item{method}{\code{method} argument of \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm}}.}

\item{name}{Character, optional name of the simulation design.}

\item{...}{Further arguments are directly stored in the list object to be passed to
\code{\link{simulate_data}}.}

\item{generator}{Function which generates data from the underlying base distribution. It is
assumend it takes the number of simulated observations \code{n_obs} as first
argument, as all random generation functions in the \pkg{stats} and
\pkg{extraDistr} do. Furthermore, it is expected to return a two-dimensional
array as output (matrix or data.frame). See details.}

\item{transform_initial}{Function which specifies the transformation of the underlying
dataset \code{Z} to final dataset \code{X}. See details.}

\item{names_final}{NULL or character vector with variable names for final dataset \code{X}.
Length needs to equal the number of columns of \code{X}. Overrides other naming options.}

\item{prefix_final}{NULL or prefix attached to variables in final dataset \code{X}. Overriden
by \code{names_final} argument. Set to NULL if no prefixes should
be added.}

\item{process_final}{List of lists specifying post-processing functions applied to final
datamatrix \code{X} before returning it. See \code{\link{process_data}}.}

\item{n_var_final}{Integer, number of columns in final datamatrix \code{X}. Can be inferred when
\code{check_and_infer} is TRUE.}

\item{types_final}{Optional vector of length equal to \code{n_var_final} (set by the user or
inferred) and hence number of columns of final dataset \code{X}.
Allowed entries are "logical", "factor" and "numeric".
Stores the type of the columns of \code{X}.
If not specified by, inferred if \code{check_and_infer} is set to TRUE.}

\item{check_and_infer}{If TRUE, then the simulation design is tested by simulating 5 observations
using \code{\link{simulate_data}}. If everything works without error,
the variables \code{n_var_final} and \code{types_final} will be inferred
from the results if not already set correctly by the user.}
}
\value{
List object with class attribute "mvtnorm_simdesign" (S3 class), inheriting
from "simdesign". It contains the same entries as a \code{\link{simdesign}} object
but in addition the following entries:

\describe{
\item{\code{mean_initial}}{}
\item{\code{sd_initial}}{}
\item{\code{cor_initial}}{Initial correlation matrix of multivariate normal distribution}
}
}
\description{
Stores information necessary to simulate and visualize datasets based
on underlying distribution multivariate normal distribution \code{Z}.
}
\details{
This S3 class implements a simulation design based on an underlying
multivariate normal distribution by creating a \code{generator} function
based on \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm}}.
}
\note{
Note that \code{relations} specifies the correlation / covariance
of the underlying gaussian data \code{Z} and thus does not directly translate into
correlations between the variables of the final datamatrix \code{X}.
}
\section{Data Generation}{

Data will be generated by \code{\link{simulate_data}} using the
following procedure:
\enumerate{
\item The underlying data matrix \code{Z} is sampled from a
multivariate gaussian distribution (number of dimensions specified by
dimensions of \code{relations}).
\item \code{Z} is then transformed into the final dataset \code{X} by applying
the \code{transform} function to \code{Z}.
\item X is post-processed if specified.
}
}

\seealso{
\code{\link{simdesign}},
\code{\link{simulate_data}},
\code{\link{simulate_data_conditional}},
\code{\link{plot_cor_network.mvtnorm_simdesign}}
}
