% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{periodicity}
\alias{periodicity}
\title{Functional returning a periodic function}
\usage{
periodicity(x_left, x_right, h, period)
}
\arguments{
\item{x_left}{Left endpoint of the nonzero interval.}

\item{x_right}{Right endpoint of the nonzero interval.}

\item{h}{The maximum height of the oscillation.}

\item{period}{The period of the function.}
}
\description{
Generates the function
h * sin(2 * pi * x * period) I(x \in [x_left x_right])
}
\examples{
f1 <- periodicity(4, 8, 3, 2)
x <- seq(2, 10, .01)
plot(x, f1(x), type = 'l')
f2 <- periodicity(5, 9, 1, .6)
lines(x, f2(x))
plot(x, cumsum(2 + f1(x) + f2(x)), type = "l")
}

