% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{periodicity_sum}
\alias{periodicity_sum}
\title{Functional returning sum of periodic functions}
\usage{
periodicity_sum(x_left, x_right, h, period)
}
\arguments{
\item{x_left}{Left endpoints of the nonzero intervals.}

\item{x_right}{Right endpoints of the nonzero intervals.}

\item{h}{The maximum heights of the oscillations.}

\item{period}{The periods of the functions.}
}
\description{
Generates the function
\sum_{j} h[j] * sin(2 * pi * x * period[j]) I(x \in [x_left[j] x_right[j]])
}
\examples{
x <- seq(2, 10, .01)
x_left <- c(1, 5, 6)
x_right <- c(3, 9, 7)
h <- c(2, 1, 3)
period <- c(1, 5, 2)
f <- periodicity_sum(x_left, x_right, h, period)
plot(x, f(x), type = "l")
plot(x, cumsum(1 + f(x)), type = "l")
}

