% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utilities.R
\name{plot_cor_network}
\alias{plot_cor_network}
\title{Visualize fixed correlation structure as a network}
\usage{
plot_cor_network(obj, categorical_indices = NULL, decimals = 2,
  cor_cutoff = 0.1, vertex_label_prefix = "z",
  edge_width_function = function(x) x * 10, vertex.size = 12,
  margin = 0, asp = 0, vertex.color = "#fff7bc",
  vertex.frame.color = "#d95f0e", vertex.label.color = "black",
  edge.color = "#fec44f", edge.label.color = "black", seed = NULL,
  ...)
}
\arguments{
\item{obj}{Correlation matrix.}

\item{categorical_indices}{Vector of indices of variables which should be drawn as rectangles
(i.e. represent categorical data).}

\item{decimals}{Number of decimals.}

\item{cor_cutoff}{Threshold of absolute correlation below which nodes are not considered
as connected. Useful to control complexity of drawn network.
Set to \code{NULL} to disable.}

\item{vertex_label_prefix}{String which is added as prefix to node labels.}

\item{edge_width_function}{Function which takes one vector input (absolute correlation values) and
outputs transformation of this vector (must be >= 0). Defines edge widths.}

\item{seed}{Set random seed to ensure reproducibility of results. Can be fixed to
obtain same layout but vary edge widths, correlation functions etc. Can
also be used to obtain nicer looking graph layouts.}
}
\description{
Useful to visualize e.g. the associations of the initial multivariate
gaussian distribution used by \code{\link{design}} and
\code{\link{simulate_data}}.
}
\details{
For plotting arguments see \code{igraph::plot}. Arguments via ... are
passed directly to \code{igraph::plot}.
}
\seealso{
\code{\link{plot_initial_cor_network}},
\code{\link{plot_transformed_cor_network}}
}
