% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot_data}
\alias{plot_data}
\title{Plots 2-dimensional data using some preset parameters.}
\usage{
plot_data(x, ...)
}
\arguments{
\item{x}{A 2-column data frame or matrix containing the data}

\item{...}{Other parameters to be passed along to \code{plot()}}
}
\description{
Plots 2-dimensional data using some preset parameters.
}

