% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utilities.R
\name{plot_estimated_cor_network}
\alias{plot_estimated_cor_network}
\title{Visualize estimated correlation matrix as a network}
\usage{
plot_estimated_cor_network(obj, n_obs = 1e+05, cor_type = "p",
  seed = NULL, show_categorical = TRUE, ...)
}
\arguments{
\item{obj}{S3 class object of type \code{simdesign} (or inheriting from it).}

\item{n_obs}{Number of simulated observations.}

\item{cor_type}{Either a character ("p" for Pearson correlation, "s" for Spearman
correlation) or a user specified function, taking one input (a data.frame)
and returning a matrix.}

\item{show_categorical}{If TRUE, marks categorical variables differently from numeric ones.
Determined by the \code{types_final} slot of the \code{obj} argument.}

\item{...}{Passed to \code{\link{plot_cor_network}}.}
}
\description{
Based on approximation via simulation specified by given simulation design.
}
\details{
This function is useful to estimate the correlation network of a simulation
setup after the initial underlying distribution \code{Z} has been transformed to
the final dataset \code{X}.
}
\seealso{
\code{\link{plot_cor_network}}
}
