% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utilities.R
\name{plot_final_cor_network}
\alias{plot_final_cor_network}
\title{Visualize estimated correlation of final simulated datamatrix}
\usage{
plot_final_cor_network(obj, n_obs = 1e+05, cor_type = "p",
  seed = NULL, show_categorical = TRUE, ...)
}
\arguments{
\item{obj}{\code{design} S3 class object.}

\item{n_obs}{Number of simulated observations.}

\item{cor_type}{Either a character ("p" for Pearson correlation, "s" for Spearman
correlation) or a user specified function, taking one input (a data.frame)
and returning a matrix.}

\item{show_categorical}{If TRUE, plots categorical variables specially. Determined by
\code{transform_type} slot of \code{design} object.}
}
\description{
Based on approximation via simulation of the final datamatrix specified by
given simulation design.
}
\details{
For explanation of further parameters, see \code{\link{plot_cor_network}}.
}
\seealso{
\code{\link{plot_cor_network}}
}
