% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utilities.R
\name{plot_initial_cor_network}
\alias{plot_initial_cor_network}
\title{Visualize correlation of initial
multivariate gaussian distribution of simulation design}
\usage{
plot_initial_cor_network(obj, ...)
}
\arguments{
\item{obj}{\code{design} S3 class object.}
}
\description{
Visualize correlation of initial
multivariate gaussian distribution of simulation design
}
\details{
For explanation of further parameters, see \code{\link{plot_cor_network}}.
}
\seealso{
\code{\link{plot_cor_network}}
}
