% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppca.R
\name{ppca_shared}
\alias{ppca_shared}
\title{Probabilistic PCA simulation}
\usage{
ppca_shared(opts = list())
}
\arguments{
\item{opts}{Options for the PPCA simulation. See merge_ppca_opts().}
}
\description{
Generate L tables X[[l]], so that
     X[[l]] = W[[l]] S[[l]]' + U * V' + E[[l]]
where the W's and S's are k[[l]]-dimensional latent scores / factors,
U and V are shared k_shared-dimensional scores / factors, generated from
spherical normals with variance opts$sigma times the identity, and E is
spherical normal with variance opts$sigma times the identity.

Note this is not the "true" ppca model -- there the prior is only on the
latent factors / sources, not scores / coordinates. Here we just need a way
to generate those coordinates, so we use a normal.
}
\examples{
ppca_res <- ppca_shared()
X <- do.call(cbind, ppca_res$X)
res <- princomp(X)
plot(res$scores)
}

