% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppls.R
\name{ppls_shared}
\alias{ppls_shared}
\title{Probabilistic PLS simulation}
\usage{
ppls_shared(opts = list())
}
\arguments{
\item{opts}{Options for the PPLS simulation. See merge_ppls_opts().}
}
\description{
Generate L tables X[[l]], so that
     X[[l]] = W[[l]] S[[l]]' + U[[l]] * V' + E[[l]]
where the W's and S's are k[[l]]-dimensional latent scores / factors,
U and V are shared k_shared-dimensional scores / factors, generated from
spherical normals with variance opts$sigma times the identity, and E is
spherical normal with variance opts$sigma times the identity.

Note this is not the "true" ppls model -- there the prior is only on the
latent factors / sources, not scores / coordinates. Here we just need a way
to generate those coordinates, so we use a normal.
}
\examples{
ppls_res <- ppls_shared()
X <- do.call(cbind, ppls_res$X)
res <- princomp(X)
plot(res$scores[, 1], ppls_res$params$U[, 1])
}

