% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{rperiodicity}
\alias{rperiodicity}
\title{Generate functions that are random sums of periodicities}
\usage{
rperiodicity(n, opts = list())
}
\arguments{
\item{n}{The total number of functions to return.}

\item{opts}{A list of options for how to generate the periodicity parameters.
See merge_rperiodicity_opts() for possible arguments.}
}
\value{
A list of functions with random parameters, which can be evaluated
on new x's.
}
\examples{
x <- seq(0, 10, .01)
f_list <- rperiodicity(10, list(M = 10))
fx <- sapply(f_list, function(f) { f(x) })
library("reshape2")
library("ggplot2")
mx <- melt(data.frame(x = x, fx = fx), id.vars = "x")
ggplot(mx) +
  geom_line(aes(x = x, y = value)) +
  facet_wrap(~variable) +
   theme_bw()
}

