% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{rsymm_spike}
\alias{rsymm_spike}
\title{Generate a list of random spike sum functions}
\usage{
rsymm_spike(n, opts = list())
}
\arguments{
\item{n}{The number of random spike sum functions to return}

\item{opts}{A list specifying how the spike functions are created. See
merge_rsymm_opts() for available options.}
}
\examples{
x <- seq(-4, 10, .01)
f_list <- rsymm_spike(10, list(M = 10)
fx <- sapply(f_list, function(f) { f(x) })
library("reshape2")
library("ggplot2")
mx <- melt(data.frame(x = x, fx = fx), id.vars = "x")
ggplot(mx) +
  geom_line(aes(x = x, y = value)) +
  facet_wrap(~variable) +
  theme_bw()
}

