% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\alias{simulate_data.default}
\alias{simulate_data.design}
\title{Simulate design matrix}
\usage{
simulate_data(relations, ...)

\method{simulate_data}{default}(relations, n_obs, transform = NULL,
  mean_initial = rep(0, nrow(relations)), sd_initial = rep(1,
  nrow(relations)), is_correlation = TRUE, names_final = NULL,
  prefix_final = "v", process_final = list(), seed = NULL)

\method{simulate_data}{design}(relations, n_obs, seed = NULL)
}
\arguments{
\item{relations}{Correlation / Covariance matrix of the initial multivariate
gaussian distribution Z or object of class \code{\link{design}}.}

\item{n_obs}{Number of simulated observations.}

\item{transform}{List of functions. Specifies transformation of underlying
datamatrix Z to final datamatrix X. If not specified, initial
data Z is returned. If named list, \code{names(transform)} will be used
for columnnames of X. See details for more information.}

\item{mean_initial}{Vector of mean values of the initial multivariate gaussian
distribution Z. Dimension needs to correspond to dimension
of \code{relations}.}

\item{sd_initial}{Vector of standard deviations of the initial multivariate
gaussian distribution Z. Dimension needs to correspond to dimension
of \code{relations}.}

\item{is_correlation}{If TRUE, then \code{relations} specifies a correlation matrix (default,
this type of specification is usually more natural than specifying
a covariance matrix). Otherwise, \code{relations} specifies a
covariance matrix.}

\item{names_final}{Variable names for final datamatrix X. Needs to have same
dimension as \code{transform}. Overrides other naming options.}

\item{prefix_final}{Prefix attached to variables in final datamatrix X. Overriden
by \code{names} argument or using a named list for \code{transform}.}

\item{process_final}{List of lists specifying post-processing functions applied to final
datamatrix X before returning it. See \code{\link{process_data}}.}

\item{seed}{Set random seed to ensure reproducibility of results.}
}
\value{
Data.frame with simulated datamatrix X.
}
\description{
Generate simulated dataset based on transformation of
multivariate gaussian distribution.
}
\details{
Data is generated using the following procedure:
\enumerate{
\item The underlying data matrix Z is sampled from a
multivariate gaussian distribution (number of dimensions specified by
dimensions of \code{relations}).
\item Z is then transformed into the final data matrix X by applying
functions from \code{transform} to the columns of Z (final number of
variables given by length of \code{transform}).
\item X is post-processed if specified (truncation to avoid
outliers).
}

Transformations are specified as a list of functions, which take
the whole datamatrix Z as single argument. For example, to multiply
column 2 of Z by 2, use function(Z) Z[,2] * 2.

Post-processing the datamatrix is based on \code{\link{process_data}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Workhorse function to be used if no \code{design}
S3 class is used.

\item \code{design}: Function to be used with \code{design} S3 class.
}}

\note{
Note that \code{relations} specifies the correlation / covariance
of the underlying gaussian data Z and thus does not directly translate into
correlations between the variables of the final datamatrix X.

This function is best used in conjunction with the \code{design} S3 class,
which facilitates further data visualization and conveniently stores
information as a template for simulation tasks.
}
\seealso{
\code{\link{design}}, \code{\link{conditional_simulate_data}},
\code{\link{process_data}}
}
