% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{symm_spike}
\alias{symm_spike}
\title{Functional for creating symmetric spikes}
\usage{
symm_spike(x0, h, w)
}
\arguments{
\item{x0}{The centerpoint for the peak.}

\item{h}{The maximum peak height.}

\item{w}{The half-width of the interval (the total width is 2 * w).}
}
\description{
Returns functions that can be evaluated on x and returns f(x)
where f is zero outside of an interval, and is a symmetric spike within that
interval.
}
\examples{
f <- symm_spike(10, 5, 3)
x <- seq(2, 20, .01)
plot(x, f(x), type = 'l')
f <- symm_spike(9, 3, 1)
lines(x, f(x))
f <- symm_spike(13, 1, 8)
lines(x, f(x))
}

