% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{symm_spike_sum}
\alias{symm_spike_sum}
\title{Functional for creating sum of symmetric spikes}
\usage{
symm_spike_sum(x0, h, w)
}
\arguments{
\item{x0}{The centerpoints for the peak.}

\item{h}{The maximum peak heights.}

\item{w}{The half-widths of the intervals.}
}
\description{
Returns functions that can be evaluated on x and returns f(x)
where f is is a sum of symmetric spike functions.
}
\examples{
x0 <- c(3, 15, 7)
h <- c(5, 3, 4)
w <- c(1, 4, .5)
f <- symm_spike_sum(x0, h, w)
x <- seq(2, 20, .01)
plot(x, f(x))
}

