% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{truncate}
\alias{truncate}
\title{Truncate columns of datamatrix}
\usage{
truncate(x, truncate_multipliers = NA)
}
\arguments{
\item{x}{Matrix or Data.frame.}

\item{truncate_multipliers}{Vector of truncation parameters. Either a single value which is
replicated as necessaary or of same dimension as \code{ncol(x)}.
If any vector entry is NA, the corresponding column will not be
truncated. See details.}
}
\value{
Matrix or Data.frame of same dimensions as input.
}
\description{
Truncate columns of datamatrix
}
\details{
Truncation is processed as follows:
\enumerate{
\item Compute the 1st and 3rd quartile q1 / q3 of transformed variables
in X.
\item Multiply these quantities by values in \code{truncate_final} to obtain
L1 and U1. If a value is NA, the corresponding variable is not truncated.
\item Set any value smaller / larger than L1 / U1 to L1 / U1.
}
}
