% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{contains_constant}
\alias{contains_constant}
\title{Check if matrix contains constant column(s)}
\usage{
contains_constant(x, eps = .Machine$double.eps)
}
\arguments{
\item{x}{Matrix or Data.frame.}

\item{eps}{Threshold for standard deviation below which a column is considered to be
constant.}
}
\value{
TRUE if one of the columns has standard deviation of below `eps``, else
FALSE.
}
\description{
Check if matrix contains constant column(s)
}
\note{
Prints a warning if constant is found.
}
