% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdesign.R
\name{discunif_simdesign}
\alias{discunif_simdesign}
\title{Uniform disc sampling design specification}
\usage{
discunif_simdesign(
  r_min = 0,
  r_max = 1,
  angle_min = 0,
  angle_max = 2 * pi,
  name = "Uniform circle simulation design",
  ...
)
}
\description{
Provides 2-dimensional points, spread uniformly over disc, or partial
disc segment.
}
\details{
The distribution of points on a disk depends on the radius - the farther out,
the more area the points need to cover. Thus, simply sampling two uniform
values for radius and angle will not work. See references.
}
\references{
\url{https://mathworld.wolfram.com/DiskPointPicking.html}
}
