% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_cor_matrix}
\alias{is_cor_matrix}
\title{Check if matrix is a correlation matrix}
\usage{
is_cor_matrix(m, tol = 1e-09)
}
\arguments{
\item{m}{Matrix.}

\item{tol}{Tolerance for checking diagonal elements.}
}
\value{
\code{TRUE} if matrix is a correlation matrix, else \code{FALSE}.
}
\description{
Checks if matrix is numeric, symmetric, has diagonal elements of one,
has only entries in \verb{[-1, 1]}, and is positive definite. Prints a warning
if a problem was found.
}
